+++ Hotel-Managment-Grp/src/Interface/ChangePass.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package hotel.managment;
+package Interface;
 
+import hotel.managment.DBConnManager;
 import java.awt.event.WindowEvent;
 import java.sql.Connection;
 import java.sql.ResultSet;
