--- Hotel-Managment-Grp/src/hotel/managment/Login.java
@@ -1,222 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package hotel.managment;
-
-
-import java.awt.Window;
-import java.awt.event.WindowEvent;
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.DriverManager;
-import java.sql.Statement;
-import java.util.Arrays;
-import javax.swing.JOptionPane;
-
-/**
- *
- * @author MaC
- */
-
-public class Login extends javax.swing.JFrame {
-
-    /**
-     * Creates new form Login
-     */
-    
-    private DBConnManager dbConnManager = null; 
-     
-    public Login() {
-        
-        
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanel1 = new javax.swing.JPanel();
-        jLabel1 = new javax.swing.JLabel();
-        txt_user = new javax.swing.JTextField();
-        jLabel2 = new javax.swing.JLabel();
-        txt_pass = new javax.swing.JPasswordField();
-        btn_login = new javax.swing.JButton();
-        btn_close = new javax.swing.JButton();
-        btn_changepass = new javax.swing.JButton();
-        jSeparator1 = new javax.swing.JSeparator();
-        jSeparator2 = new javax.swing.JSeparator();
-        jLabel4 = new javax.swing.JLabel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        getContentPane().setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        jPanel1.setBackground(new java.awt.Color(0, 0, 41));
-        jPanel1.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        jLabel1.setBackground(new java.awt.Color(255, 255, 255));
-        jLabel1.setFont(new java.awt.Font("Tahoma", 1, 12)); // NOI18N
-        jLabel1.setForeground(new java.awt.Color(102, 102, 255));
-        jLabel1.setText("Username");
-        jPanel1.add(jLabel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 90, -1, -1));
-
-        txt_user.setBackground(new java.awt.Color(0, 0, 41));
-        txt_user.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        txt_user.setForeground(new java.awt.Color(255, 255, 255));
-        txt_user.setBorder(null);
-        jPanel1.add(txt_user, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 110, 340, -1));
-
-        jLabel2.setBackground(new java.awt.Color(255, 255, 255));
-        jLabel2.setFont(new java.awt.Font("Tahoma", 1, 12)); // NOI18N
-        jLabel2.setForeground(new java.awt.Color(102, 102, 255));
-        jLabel2.setText("Password");
-        jPanel1.add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 150, -1, -1));
-
-        txt_pass.setBackground(new java.awt.Color(0, 0, 41));
-        txt_pass.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        txt_pass.setForeground(new java.awt.Color(255, 255, 255));
-        txt_pass.setBorder(null);
-        jPanel1.add(txt_pass, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 170, 340, -1));
-
-        btn_login.setBackground(new java.awt.Color(102, 102, 255));
-        btn_login.setText("Login");
-        btn_login.setBorder(null);
-        btn_login.setBorderPainted(false);
-        btn_login.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_loginActionPerformed(evt);
-            }
-        });
-        jPanel1.add(btn_login, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 250, 121, 39));
-
-        btn_close.setText("Close");
-        btn_close.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_closeActionPerformed(evt);
-            }
-        });
-        jPanel1.add(btn_close, new org.netbeans.lib.awtextra.AbsoluteConstraints(220, 250, 116, 39));
-
-        btn_changepass.setText("Change Password");
-        btn_changepass.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_changepassActionPerformed(evt);
-            }
-        });
-        jPanel1.add(btn_changepass, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 20, -1, -1));
-        jPanel1.add(jSeparator1, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 190, 340, 10));
-        jPanel1.add(jSeparator2, new org.netbeans.lib.awtextra.AbsoluteConstraints(80, 130, 340, 10));
-
-        getContentPane().add(jPanel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(240, 0, 500, 330));
-
-        jLabel4.setFont(new java.awt.Font("Tahoma", 1, 36)); // NOI18N
-        jLabel4.setText("LOGIN");
-        getContentPane().add(jLabel4, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 130, -1, -1));
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void btn_loginActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_loginActionPerformed
-        JOptionPane.showMessageDialog(null,log());
-    }//GEN-LAST:event_btn_loginActionPerformed
-
-    private void btn_closeActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_closeActionPerformed
-       Login f = new Login(); 
-       f.dispatchEvent(new WindowEvent(f, WindowEvent.WINDOW_CLOSING));
-    }//GEN-LAST:event_btn_closeActionPerformed
-
-    private void btn_changepassActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_changepassActionPerformed
-        
-       
-        ChangePass cp = new ChangePass();
-        cp.setVisible(true);
-        
-    }//GEN-LAST:event_btn_changepassActionPerformed
-
-    public String log(){
-        dbConnManager = new DBConnManager();  
-        Connection dbConn = null;
-        String result = null;
-        
-        try{
-            dbConn = dbConnManager.connect();
-            Statement stmt = dbConn.createStatement();
-            String query = "SELECT * FROM login WHERE username = '"+txt_user.getText()+"' and password = '"+String.valueOf(txt_pass.getPassword())+"' ";
-            
-            ResultSet rs = stmt.executeQuery(query);
-            if(rs.next()){
-                result = "Access Granted !";
-            }
-            else
-            {
-                result = "Access Denied!";
-            }
-        }
-        catch(SQLException ex)
-        {
-            System.out.println(ex + "-----------The entered student ID cannot be found.");  
-            
-        }finally {             
-            dbConnManager.connectionClose(dbConn);         
-        }   
-        
-        return result;
-    }
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(Login.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(Login.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(Login.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(Login.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new Login().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btn_changepass;
-    private javax.swing.JButton btn_close;
-    private javax.swing.JButton btn_login;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JSeparator jSeparator1;
-    private javax.swing.JSeparator jSeparator2;
-    private javax.swing.JPasswordField txt_pass;
-    private javax.swing.JTextField txt_user;
-    // End of variables declaration//GEN-END:variables
-}
